+++ packages/main-api/src/controllers/m2/binances/get_merchant_balance.rs
@@ -5,12 +5,9 @@ use dto::{
     by_axum::auth::Authorization,
     sqlx::{Pool, Postgres},
 };
-use hmac::{Hmac, Mac};
-use rand::{Rng, distr::Alphanumeric, rng};
-use sha2::Sha512;
 
 use crate::config;
-use crate::sqlx::types::time::OffsetDateTime;
+use crate::utils::crypto::{apply_binance_headers, sign_for_binance};
 
 #[derive(
     Debug,
@@ -23,9 +20,7 @@ use crate::sqlx::types::time::OffsetDateTime;
     JsonSchema,
 )]
 pub struct MerchantBalanceRequest {
-    #[schemars(description = "FUNDING_WALLET or SPOT_WALLET")]
     pub wallet: String,
-    #[schemars(description = "USDT")]
     pub currency: String,
 }
 
@@ -43,21 +38,6 @@ pub struct MerchantBalanceResponse {
     pub balance: f64,
 }
 
-fn sign_sha512_upper_hex(secret: &str, payload: &str) -> String {
-    let mut mac = Hmac::<Sha512>::new_from_slice(secret.as_bytes()).expect("hmac key");
-    mac.update(payload.as_bytes());
-    let sig = mac.finalize().into_bytes();
-    hex::encode_upper(sig)
-}
-
-fn gen_nonce_32_alnum() -> String {
-    rng()
-        .sample_iter(&Alphanumeric)
-        .take(32)
-        .map(char::from)
-        .collect()
-}
-
 pub async fn binance_merchant_balance_handler(
     Extension(_auth): Extension<Option<Authorization>>,
     State(_pool): State<Pool<Postgres>>,
@@ -69,34 +49,18 @@ pub async fn binance_merchant_balance_handler(
     let api_key = conf.binance.api_key;
     let secret = conf.binance.secret_key;
 
-    let body_json = serde_json::json!({
+    let body_value = serde_json::json!({
         "wallet": req.wallet,
         "currency": req.currency
-    })
-    .to_string();
+    });
 
-    let ts = (OffsetDateTime::now_utc().unix_timestamp_nanos() / 1_000_000).to_string();
-    let nonce = gen_nonce_32_alnum();
-
-    let sign_payload = format!(
-        "{ts}\n{nonce}\n{body}\n",
-        ts = ts,
-        nonce = nonce,
-        body = body_json
-    );
-    let signature = sign_sha512_upper_hex(&secret, &sign_payload);
+    let (ts, nonce, signature) = sign_for_binance(&secret, &body_value)?;
 
     let client = reqwest::Client::new();
     let url = format!("{}/v2/balance", base);
 
-    let resp = client
-        .post(url)
-        .header("content-type", "application/json")
-        .header("BinancePay-Timestamp", &ts)
-        .header("BinancePay-Nonce", &nonce)
-        .header("BinancePay-Certificate-SN", api_key)
-        .header("BinancePay-Signature", &signature)
-        .body(body_json)
+    let resp = apply_binance_headers(client.post(url), &api_key, &ts, &nonce, &signature)
+        .body(body_value.to_string())
         .send()
         .await
         .map_err(|e| dto::Error::ServerError(e.to_string()))?;
@@ -111,7 +75,6 @@ pub async fn binance_merchant_balance_handler(
         serde_json::from_str(&text).unwrap_or_else(|_| serde_json::json!({ "raw": text }));
 
     let ok = status.is_success() && json.get("status").and_then(|v| v.as_str()) == Some("SUCCESS");
-
     if !ok {
         let code = json
             .get("code")
@@ -127,7 +90,6 @@ pub async fn binance_merchant_balance_handler(
     }
 
     let data = &json["data"];
-    tracing::debug!("balance data: {:?}", data);
 
     fn as_f64_any(v: &serde_json::Value) -> Option<f64> {
         v.as_f64()
