+++ packages/main-api/src/controllers/v2/binances/binance_webhook.rs
@@ -1,5 +1,5 @@
-use base64::Engine as _;
-use base64::engine::general_purpose::STANDARD as B64;
+use crate::config;
+use crate::utils::crypto::verify_webhook_signature;
 use bdk::prelude::*;
 use by_axum::axum::{Extension, Json, body::Bytes, extract::State, http::HeaderMap};
 use dto::{
@@ -8,10 +8,7 @@ use dto::{
     by_axum::auth::Authorization,
     sqlx::{Pool, Postgres},
 };
-use openssl::{hash::MessageDigest, pkey::PKey, sign::Verifier};
 use serde_json::Value;
-
-use crate::{config, utils::wallets::sign_for_binance::sign_for_binance};
 
 pub async fn binance_webhook_handler(
     Extension(_auth): Extension<Option<Authorization>>,
@@ -45,51 +42,11 @@ pub async fn binance_webhook_handler(
     );
 
     let raw_body = String::from_utf8_lossy(&body);
-    let payload = format!("{ts}\n{nonce}\n{raw_body}\n");
-    tracing::debug!("payload composed ({} bytes)", payload.len());
 
-    let certs = fetch_all_cert_pems(&base, &api_key, &secret)
-        .await
-        .map_err(|e| dto::Error::ServerError(format!("certificates failed: {e}")))?;
-    tracing::debug!("fetched {} cert(s)", certs.len());
-
-    let mut tried = 0usize;
-    let mut verified = false;
-
-    if let Some(sn) = cert_sn {
-        if let Some(pem) = certs.iter().find_map(|(this_sn, pem)| {
-            if this_sn == sn {
-                Some(pem.as_str())
-            } else {
-                None
-            }
-        }) {
-            tried += 1;
-            if verify_rsa_sha256(pem, payload.as_bytes(), sig_b64).unwrap_or(false) {
-                verified = true;
-            } else {
-                tracing::warn!("verify failed with cert_sn={}", sn);
-            }
-        } else {
-            tracing::warn!("cert_sn from header not found in /certificates: {}", sn);
-        }
-    }
-
-    if !verified {
-        for (this_sn, pem) in &certs {
-            tried += 1;
-            if verify_rsa_sha256(pem, payload.as_bytes(), sig_b64).unwrap_or(false) {
-                tracing::debug!("verify success with cert_sn={}", this_sn);
-                verified = true;
-                break;
-            }
-        }
-    }
-
-    tracing::debug!("verify result: {}, tried {} cert(s)", verified, tried);
-    if !verified {
-        return Err(dto::Error::Unauthorized);
-    }
+    verify_webhook_signature(
+        &base, &api_key, &secret, ts, nonce, sig_b64, cert_sn, &raw_body,
+    )
+    .await?;
 
     let biz_status = parse_biz_status(&raw_body).unwrap_or_default();
     tracing::debug!("biz status: {:?}", biz_status);
@@ -175,95 +132,6 @@ fn get_header<'a>(headers: &'a HeaderMap, key: &str) -> Result<&'a str> {
         .get(key)
         .and_then(|v| v.to_str().ok())
         .ok_or_else(|| dto::Error::ServerError(format!("missing header: {key}")))
-}
-
-async fn fetch_all_cert_pems(
-    base: &str,
-    api_key: &str,
-    secret: &str,
-) -> std::result::Result<Vec<(String, String)>, String> {
-    let body = serde_json::json!({});
-    let (ts, nonce, sign) =
-        sign_for_binance(secret, &body).map_err(|e| format!("sign_for_binance failed: {e}"))?;
-    let url = format!("{}/certificates", base);
-
-    let client = reqwest::Client::new();
-    let resp = client
-        .post(url)
-        .header("content-type", "application/json")
-        .header("BinancePay-Timestamp", &ts)
-        .header("BinancePay-Nonce", &nonce)
-        .header("BinancePay-Certificate-SN", api_key)
-        .header("BinancePay-Signature", &sign)
-        .body(body.to_string())
-        .send()
-        .await
-        .map_err(|e| format!("request failed: {e}"))?;
-
-    let status = resp.status();
-    let text = resp
-        .text()
-        .await
-        .map_err(|e| format!("read body failed: {e}"))?;
-    tracing::debug!("certificates http={} body={}", status, text);
-
-    let v: Value =
-        serde_json::from_str(&text).unwrap_or_else(|_| serde_json::json!({ "raw": text }));
-    if !status.is_success() || v.get("status").and_then(|s| s.as_str()) != Some("SUCCESS") {
-        let code = v.get("code").and_then(|x| x.as_str()).unwrap_or("UNKNOWN");
-        let msg = v
-            .get("errorMessage")
-            .and_then(|x| x.as_str())
-            .unwrap_or("no errorMessage");
-        return Err(format!(
-            "certificates api failed: http={status}, code={code}, msg={msg}"
-        ));
-    }
-
-    let arr = v
-        .get("data")
-        .and_then(|d| d.as_array())
-        .ok_or("no data array")?;
-    let mut out = Vec::with_capacity(arr.len());
-    for it in arr {
-        let sn = it
-            .get("certSn")
-            .or_else(|| it.get("certSerial"))
-            .and_then(|x| x.as_str())
-            .ok_or("certSn/certSerial missing")?;
-
-        let pem = it
-            .get("certPublic")
-            .and_then(|x| x.as_str())
-            .ok_or("certPublic missing")?;
-
-        if PKey::public_key_from_pem(pem.as_bytes()).is_ok() {
-            out.push((sn.to_string(), pem.to_string()));
-        } else {
-            tracing::warn!("invalid PUBLIC KEY PEM for cert serial={}", sn);
-        }
-    }
-    Ok(out)
-}
-
-fn verify_rsa_sha256(
-    cert_pem: &str,
-    payload: &[u8],
-    sig_b64: &str,
-) -> std::result::Result<bool, String> {
-    let pkey = PKey::public_key_from_pem(cert_pem.as_bytes())
-        .map_err(|e| format!("load pem failed: {e}"))?;
-    let sig = B64
-        .decode(sig_b64.as_bytes())
-        .map_err(|e| format!("b64 decode failed: {e}"))?;
-    let mut verifier = Verifier::new(MessageDigest::sha256(), &pkey)
-        .map_err(|e| format!("verifier init failed: {e}"))?;
-    verifier
-        .update(payload)
-        .map_err(|e| format!("verifier update failed: {e}"))?;
-    verifier
-        .verify(&sig)
-        .map_err(|e| format!("verifier verify failed: {e}"))
 }
 
 fn parse_biz_status(raw_body: &str) -> Option<String> {
