+++ packages/main-api/src/controllers/v2/binances/unsubscribe.rs
@@ -10,8 +10,8 @@ use serde_json::Value;
 use crate::{
     config,
     utils::{
-        generate_merchant_trade_no::gen_merchant_trade_no, users::extract_user,
-        wallets::sign_for_binance::sign_for_binance,
+        crypto::sign_for_binance, generate_merchant_trade_no::gen_merchant_trade_no,
+        users::extract_user,
     },
 };
