+++ packages/main-api/src/utils/wallets/mod.rs
@@ -1,7 +1,6 @@
 pub mod kaia_local_wallet;
 pub mod kaia_transaction;
 pub mod local_fee_payer;
-pub mod sign_for_binance;
 
 use async_trait::async_trait;
 use ethers::types::{H160, Signature};
